+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 ## [0.32.0] - xxx
+### Added
+- Implemented ```date-format```
 ### Fixed
 - [```args-only```](https://github.com/frosklis/dinero-rs/issues/120)
+
 ## [0.31.0] - 2021-08-22 
 ### Added
 - [The ```roi``` command](https://github.com/frosklis/dinero-rs/issues/115) is good enough